@@ -1,53 +1,51 @@
 {
   "name": "pas-frontend",
   "version": "0.1.0",
-  "description": "PAS Map Front End Client",
+  "description": "PAS Map UI",
   "author": "Chris Wade <genghishack@gmail.com>",
   "repository": "https://github.com/genghishack/pas-frontend.git",
   "license": "GPL-3.0-or-later",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.34",
-    "@fortawesome/free-solid-svg-icons": "^5.15.2",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@mapbox/geo-viewport": "^0.4.1",
-    "@material-ui/core": "^4.11.3",
-    "@reduxjs/toolkit": "^1.1.0",
-    "@types/react-bootstrap": "^0.32.25",
-    "@types/react-router-bootstrap": "^0.24.5",
-    "aws-amplify": "^3.3.14",
-    "leaflet": "^1.7.1",
-    "moment": "^2.29.1",
-    "rc-slider": "^9.6.4",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.4.3",
+    "@fortawesome/fontawesome-svg-core": "^6.4.2",
+    "@fortawesome/free-solid-svg-icons": "^6.4.2",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@mapbox/geo-viewport": "^0.5.0",
+    "@reduxjs/toolkit": "^1.9.7",
+    "aws-amplify": "^5.3.11",
+    "leaflet": "^1.9.4",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.9.1",
     "react-context-devtool": "^2.0.3",
-    "react-dom": "^17.0.1",
-    "react-icons": "^4.1.0",
-    "react-leaflet": "^3.0.5",
-    "react-redux": "^7.1.3",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.2.0"
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.11.0",
+    "react-leaflet": "^4.2.1",
+    "react-redux": "^8.1.3",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.17.0"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^26.0.15",
-    "@types/leaflet": "^1.5.21",
-    "@types/mapbox__geo-viewport": "^0.4.1",
-    "@types/node": "^14.14.6",
-    "@types/react": "^16.9.56",
-    "@types/react-dom": "^16.9.9",
-    "@types/react-redux": "^7.1.11",
-    "@types/react-router-dom": "^5.1.6",
-    "@typescript-eslint/parser": "^4.6.1",
-    "eslint": "^7.13.0",
-    "node-sass": "4.14.1",
-    "react-scripts": "4.0.0",
-    "source-map-loader": "^1.1.2",
-    "ts-loader": "^8.0.10",
-    "typescript": "~4.0.5"
+    "@testing-library/react": "^14.0.0",
+    "@types/leaflet": "^1.9.7",
+    "@types/mapbox__geo-viewport": "^0.5.2",
+    "@types/node": "^20.8.9",
+    "@types/react": "^18.2.33",
+    "@types/react-bootstrap": "^0.32.34",
+    "@types/react-dom": "^18.2.14",
+    "@types/react-redux": "^7.1.28",
+    "@types/react-router-bootstrap": "^0.26.5",
+    "@types/react-router-dom": "^5.3.3",
+    "@typescript-eslint/parser": "^5.62.0",
+    "eslint": "^8.52.0",
+    "node-sass": "^7.0.3",
+    "react-scripts": "^5.0.1",
+    "source-map-loader": "^4.0.1",
+    "ts-loader": "^9.5.0",
+    "typescript": "^5.2.2"
+  },
+  "overrides": {
+    "typescript": "^5.2.2"
   },
   "scripts": {
     "start": "react-scripts start",
